requirements.txt
# ByBit API
pybit>=5.7.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-socketio>=5.10.0

# Telegram Bot
python-telegram-bot>=20.7

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Environment variables
python-dotenv>=1.0.0

# Async HTTP client
httpx>=0.25.0

# Logging and utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0

